更新com.blankj:utilcode  && 修改 Account成员变量属性

+++ app/src/main/java/shellhub/github/neteasemusic/response/login/Account.java
@@ -43,7 +43,7 @@ public class Account implements Serializable {
 	private int baoyueVersion;
 
 	@SerializedName("viptypeVersion")
-	private int viptypeVersion;
+	private double viptypeVersion;
 
 	@SerializedName("donateVersion")
 	private int donateVersion;
